+++ Zwaeg/Features/Scanner/NutritionLabelParser.swift
@@ -122,7 +122,7 @@ enum NutritionLabelParser {
     }
 }
 
-private extension CGImagePropertyOrientation {
+extension CGImagePropertyOrientation {
     init(_ orientation: UIImage.Orientation) {
         switch orientation {
         case .up: self = .up